fix: remove azure claim overage code.
