+++ internal/api/provider/oidc.go
@@ -65,7 +65,7 @@ func ParseIDToken(ctx context.Context, provider *oidc.Provider, config *oidc.Con
 		token, data, err = parseVercelMarketplaceIDToken(token)
 	default:
 		if IsAzureIssuer(token.Issuer) {
-			token, data, err = parseAzureIDToken(ctx, token, options.AccessToken)
+			token, data, err = parseAzureIDToken(token)
 		} else {
 			token, data, err = parseGenericIDToken(token)
 		}
@@ -206,6 +206,111 @@ func parseLinkedinIDToken(token *oidc.IDToken) (*oidc.IDToken, *UserProvidedData
 		Locale:     claims.Locale,
 		Picture:    claims.Picture,
 		ProviderId: token.Subject,
+	}
+
+	return token, &data, nil
+}
+
+type AzureIDTokenClaims struct {
+	jwt.RegisteredClaims
+
+	Email                              string `json:"email"`
+	Name                               string `json:"name"`
+	PreferredUsername                  string `json:"preferred_username"`
+	XMicrosoftEmailDomainOwnerVerified any    `json:"xms_edov"`
+}
+
+func (c *AzureIDTokenClaims) IsEmailVerified() bool {
+	emailVerified := false
+
+	edov := c.XMicrosoftEmailDomainOwnerVerified
+
+	// If xms_edov is not set, and an email is present or xms_edov is true,
+	// only then is the email regarded as verified.
+	// https://learn.microsoft.com/en-us/azure/active-directory/develop/migrate-off-email-claim-authorization#using-the-xms_edov-optional-claim-to-determine-email-verification-status-and-migrate-users
+	if edov == nil {
+		// An email is provided, but xms_edov is not -- probably not
+		// configured, so we must assume the email is verified as Azure
+		// will only send out a potentially unverified email address in
+		// single-tenanat apps.
+		emailVerified = c.Email != ""
+	} else {
+		edovBool := false
+
+		// Azure can't be trusted with how they encode the xms_edov
+		// claim. Sometimes it's "xms_edov": "1", sometimes "xms_edov": true.
+		switch v := edov.(type) {
+		case bool:
+			edovBool = v
+
+		case string:
+			edovBool = v == "1" || v == "true"
+
+		default:
+			edovBool = false
+		}
+
+		emailVerified = c.Email != "" && edovBool
+	}
+
+	return emailVerified
+}
+
+// removeAzureClaimsFromCustomClaims contains the list of claims to be removed
+// from the CustomClaims map. See:
+// https://learn.microsoft.com/en-us/azure/active-directory/develop/id-token-claims-reference
+var removeAzureClaimsFromCustomClaims = []string{
+	"aud",
+	"iss",
+	"iat",
+	"nbf",
+	"exp",
+	"c_hash",
+	"at_hash",
+	"aio",
+	"nonce",
+	"rh",
+	"uti",
+	"jti",
+	"ver",
+	"sub",
+	"name",
+	"preferred_username",
+}
+
+func parseAzureIDToken(token *oidc.IDToken) (*oidc.IDToken, *UserProvidedData, error) {
+	var data UserProvidedData
+
+	var azureClaims AzureIDTokenClaims
+	if err := token.Claims(&azureClaims); err != nil {
+		return nil, nil, err
+	}
+
+	data.Metadata = &Claims{
+		Issuer:            token.Issuer,
+		Subject:           token.Subject,
+		ProviderId:        token.Subject,
+		PreferredUsername: azureClaims.PreferredUsername,
+		FullName:          azureClaims.Name,
+		CustomClaims:      make(map[string]any),
+	}
+
+	if azureClaims.Email != "" {
+		data.Emails = []Email{{
+			Email:    azureClaims.Email,
+			Verified: azureClaims.IsEmailVerified(),
+			Primary:  true,
+		}}
+	}
+
+	if err := token.Claims(&data.Metadata.CustomClaims); err != nil {
+		return nil, nil, err
+	}
+
+	if data.Metadata.CustomClaims != nil {
+		for _, claim := range removeAzureClaimsFromCustomClaims {
+			delete(data.Metadata.CustomClaims, claim)
+		}
 	}
 
 	return token, &data, nil
